Docstrings (#112)
* Move path classes to new file

* Put obstacle classes in their own file

* Clean method

* Clean method

* Fix dependency issue

* Clean Path.py

* Clean code

* Clean code

* Save progress

* Fix changes

* Save code

* Add in last docstrings

* Clean code

* Document Path.py
